+++ src/main/scala/com/raquo/laminar/lifecycle/MountContext.scala
@@ -1,0 +1,5 @@
+package com.raquo.laminar.lifecycle
+
+class MountContext {
+
+}