+++ CHANGELOG.md
@@ -4,6 +4,11 @@ Breaking changes in **bold**.
 
 #### v0.8 – TBD
 
+* Mount Events renamed to Lifecycle Events
+  - mountEvents -> lifecycleEvents
+  - maybeThisLifecycleMountEventBus
+  - MountEvent -> LifecycleEvent
+  - New: mountEvents / unmountEvents
 * **API: Lifecycle events & Ownership overhaul**
   * <TODO: Elaborate> Use Airstream's new `DynamicOwner` and `DynamicSubscription` features: `ReactiveElement` subscriptions are now activated only when the element is mounted, not when it is created.
   * <TODO: Elaborate> Transaction delay for `mountEvents` now applies to all element subscriptions